+++ huey/contrib/sql_huey.py
@@ -218,6 +218,16 @@ class SqlStorage(BaseStorage):
         else:
             return kv.value
 
+    def peek_many(self, keys):
+        self.check_conn()
+        accum = {}
+        for i in range(0, len(keys), 500):
+            query = (self.kv(self.KV.key, self.KV.value)
+                     .where(self.KV.key.in_(keys[i:i + 500]))
+                     .tuples())
+            accum.update(query)
+        return accum
+
     def pop_data(self, key):
         self.check_conn()
         query = self.kv().where(self.KV.key == key)
@@ -239,7 +249,9 @@ class SqlStorage(BaseStorage):
         self.check_conn()
         return self.kv().where(self.KV.key == key).exists()
 
-    def put_if_empty(self, key, value):
+    def put_if_empty(self, key, value, ttl=None):
+        if ttl is not None:
+            raise NotImplementedError('ttl is not supported by this storage.')
         self.check_conn()
         try:
             with self.database.atomic():